Add alternative path to delete the contents of a volume in the host.
