+++ pkg/util/csi_proxy_v1_windows.go
@@ -117,20 +117,6 @@ func (proxy *CSIProxyV1) IsSymlink(mountPath string) (isSymlink bool, err error)
 	return isSymlinkResponse.IsSymlink, nil
 }
 
-// Lsdir lists files under a directory.
-func (proxy *CSIProxyV1) Lsdir(path string) (files []string, err error) {
-	lsdirResponse, err := proxy.FilesystemClient.Lsdir(
-		context.Background(),
-		&filesystemapi.LsdirRequest{
-			Path: path,
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-	return lsdirResponse.Files, nil
-}
-
 // Mkdir makes a directory.
 func (proxy *CSIProxyV1) Mkdir(path string) error {
 	_, err := proxy.FilesystemClient.Mkdir(
