+++ pkg/util/volume_util.go
@@ -59,3 +59,17 @@ func (u *volumeUtil) IsDir(fullPath string) (bool, error) {
 
 	return stat.IsDir(), nil
 }
+
+// ReadDir returns a list all the files under the given directory
+func (u *volumeUtil) ReadDir(fullPath string) ([]string, error) {
+	dir, err := os.Open(fullPath)
+	if err != nil {
+		return nil, err
+	}
+	defer dir.Close()
+	files, err := dir.Readdirnames(-1)
+	if err != nil {
+		return nil, err
+	}
+	return files, nil
+}
