+++ pkg/util/volume_util_linux.go
@@ -109,18 +109,3 @@ func (u *volumeUtil) DeleteContents(hostPath, mountPath string) error {
 
 	return utilerrors.NewAggregate(errList)
 }
-
-
-// ReadDir returns a list all the files under the given directory
-func (u *volumeUtil) ReadDir(fullPath string) ([]string, error) {
-	dir, err := os.Open(fullPath)
-	if err != nil {
-		return nil, err
-	}
-	defer dir.Close()
-	files, err := dir.Readdirnames(-1)
-	if err != nil {
-		return nil, err
-	}
-	return files, nil
-}
