+++ pkg/util/volume_util_windows.go
@@ -23,7 +23,8 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
+
+	"k8s.io/klog/v2"
 )
 
 var _ VolumeUtil = &volumeUtil{}
@@ -56,18 +57,28 @@ func (u *volumeUtil) GetFsCapacityByte(hostPath, mountPath string) (int64, error
 		return 0, fmt.Errorf("file mountPath=%q is not a symlink", mountPath)
 	}
 
-	// symlinkTarget is the result of dereferencing the symlink
-	// it could be in the form Volume{<volumeid>} or an actual file
-	// outside the mounted dir
-	symlinkTarget, err := os.Readlink(mountPath)
-	if err != nil {
-		return 0, err
-	}
+	var volumeID string
+	var errNoVolumeID, errNoMoveUpVolumeID error
 
-	// initially assume that the symlink points to a Volume{<volumeid>},
-	// the path to the volume is the hostPath
+	// initially assume that the symlink points to a Volume{<volumeid>}
 	volumePath := hostPath
-	if !strings.HasPrefix(symlinkTarget, "Volume{") {
+	volumeID, errNoVolumeID = u.csiProxy.GetVolumeId(volumePath)
+
+	if errNoVolumeID != nil {
+		klog.V(5).Infof("path=%s is not a Volume, attempting to follow the symlink to get the volume capacity", volumePath)
+
+		// in the case where it isn't a Volume{<volumeid>} assume that
+		// after dereferencing the symlink and moving one level up
+		// in the hierarchy we get a volumeID
+		//
+		// symlinkTarget is the result of dereferencing the symlink
+		// it could be in the form Volume{<volumeid>} or an actual file
+		// outside the mounted dir
+		symlinkTarget, err := os.Readlink(mountPath)
+		if err != nil {
+			return 0, err
+		}
+
 		// the symlink is pointing to a directory, assume that it has this structure
 		//
 		// volume/ (symlink -> \\\Volume{}\)
@@ -77,12 +88,12 @@ func (u *volumeUtil) GetFsCapacityByte(hostPath, mountPath string) (int64, error
 		//
 		// symlinkTarget is volume/dir0 at this point, trim the last element of the path
 		volumePath = filepath.Dir(symlinkTarget)
+		volumeID, errNoMoveUpVolumeID = u.csiProxy.GetVolumeId(volumePath)
+		if errNoMoveUpVolumeID != nil {
+			return 0, fmt.Errorf("Failed to find a volume either by checking if file=%s is a volume or by following the symlink, moving one level up and checking if file=%s is a volume", hostPath, volumePath, errNoMoveUpVolumeID)
+		}
 	}
 
-	volumeID, err := u.csiProxy.GetVolumeId(volumePath)
-	if err != nil {
-		return 0, err
-	}
 	totalBytes, _, err := u.csiProxy.GetVolumeStats(volumeID)
 	if err != nil {
 		return 0, err
@@ -102,27 +113,17 @@ func (u *volumeUtil) recreateDirectory(hostPath string) error {
 	return nil
 }
 
-// DeleteContents deletes all the contents under the given directory
+// DeleteContents deletes all the contents under the given directory.
 func (u *volumeUtil) DeleteContents(hostPath, mountPath string) error {
 	// mountPath is in the context of the volume inside local volume provisioner
 	// the path to use in Windows is the one that CSI Proxy will use and it should
 	// be in the context of the host (because CSI Proxy doesn't know about the context
 	// of the local volume provisioner volumes)
 
-	// symlinkTarget is the result of dereferencing the symlink
-	// it could be in the form Volume{<volumeid>} or an actual file
-	// outside the mounted dir
-	symlinkTarget, err := os.Readlink(mountPath)
-	if err != nil {
-		return err
-	}
-
-	if strings.HasPrefix(symlinkTarget, "Volume{") {
-		// it's pointing to a Volume
-		volumeID, err := u.csiProxy.GetVolumeId(hostPath)
-		if err != nil {
-			return err
-		}
+	// assume that it's pointing to a Volume
+	volumeID, err := u.csiProxy.GetVolumeId(hostPath)
+	if err == nil {
+		// if so, reformat the volume
 		err = u.csiProxy.FormatVolume(volumeID)
 		if err != nil {
 			return err
@@ -130,7 +131,18 @@ func (u *volumeUtil) DeleteContents(hostPath, mountPath string) error {
 		return nil
 	}
 
-	// it's pointing to a directory in the host
+	// otherwise it's pointing to a directory in the host
+	klog.V(5).Info("Cannot format hostPath=%q because it's not a Volume, attempting to follow the symlink", hostPath)
+
+	// symlinkTarget is the result of dereferencing the symlink
+	// this path only makes sense in the context of the host
+	symlinkTarget, err := os.Readlink(mountPath)
+	if err != nil {
+		return err
+	}
+
+	klog.V(5).Info("Attempting to delete hostPath=%q through CSI Proxy", symlinkTarget)
+	// recreate the directory that the symlink points to
 	err = u.recreateDirectory(symlinkTarget)
 	if err != nil {
 		return err
@@ -158,13 +170,4 @@ func (u *volumeUtil) IsLikelyMountPoint(hostPath, mountPath string, mountPointMa
 		return false, fmt.Errorf("hostPath %q is not a symlink", hostPath)
 	}
 	return isLikelyMountPoint, nil
-}
-
-// ReadDir returns a list all the files under the given directory
-func (u *volumeUtil) ReadDir(fullPath string) ([]string, error) {
-	files, err := u.csiProxy.Lsdir(fullPath)
-	if err != nil {
-		return nil, err
-	}
-	return files, nil
 }
